requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
django-filter
python-decouple
drf-yasg
celery
redis